refactor(plugins): Extract registration of extensions to its own file. - add tests. - move the IDeckPlugin interface - remove IPluginManifest interface
