+++ app/scripts/modules/core/src/plugins/index.ts
@@ -1,2 +1,3 @@
+export * from './deck.plugin';
 export * from './plugin.module';
 export * from './plugin.registry';
